pyproject.toml
[project]
name = "ai-document-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.121.1",
    "langchain>=1.0.5",
    "langchain-openai>=1.0.2",
    "python-dotenv>=1.2.1",
    "python-json-logger>=4.0.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
